Move XmlException into a separate file
